+++ packages/polaris-viz/CHANGELOG.md
@@ -20,6 +20,9 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 - Renamed `Annotation.tooltipData.label` to `key`.
 - Renamed `theme.tooltip.valueColor` to `textColor`.
 - Renamed `theme.tooltip.labelColor` to `titleColor`.
+- Renamed `theme.tooltip.labelColor` to `titleColor`.
+- Changed `LegendData.iconType` to `shape` which now accepts the `Shape` type.
+- Changed `useLegend()` `type` to `shape` which now accepts the `Shape` typ
 
 ### Removed